pyproject.toml
[tool.poetry]
name = "news"
version = "0.0.1"
description = "News editor"
authors = ["Yoel Benítez Fonseca <ybenitezf@gmail.com>"]
license = "MIT-3"
readme = "README.md"

packages = [
    {include = "development"},
    {include = "news/ui_back",from = "bases"},
]

[tool.poetry.dependencies]
python = "^3.10"
python-fasthtml = "^0.12.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
ruff = "^0.8.1"
ruff-lsp = "^0.0.59"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"

[tool.poetry.group.typing.dependencies]
mypy = "^1.13.0"

[tool.ruff]
line-length = 79
target-version = "py311"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "W", "N"]
ignore = []

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 79

[tool.mypy]
mypy_path = "components, bases"
namespace_packages = true
explicit_package_bases = true
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_calls = true
exclude = "(?x)(^test/|^projects/)"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
